Move CfgMagazines to dedicated .hpp file
