+++ addons/main/CfgMagazines.hpp
@@ -1,0 +1,11 @@
+class CfgMagazines {	
+	class CA_Magazine;
+	class DRNP_AR2_battery: CA_Magazine {
+		picture = "\z\DRNP\addons\main\UI\Icon_Batt.paa";
+		author = "Pastoman1";
+		scope = 2;
+		displayName = "Portable drone battery";
+		ammo = "Ar2_battery";
+		count = 100;
+	};
+};
