+++ addons/main/CfgVehicles.hpp
