+++ addons/main/config.cpp
@@ -59,18 +59,8 @@ class Extended_Init_EventHandlers {
 };
 
 class CBA_Extended_EventHandlers_base;
-class CfgMagazines {	
-	class CA_Magazine;
-	class DRNP_AR2_battery: CA_Magazine {
-		picture = "\z\DRNP\addons\main\UI\Icon_Batt.paa";
-		author = "Pastoman1";
-		scope = 2;
-		displayName = "Portable drone battery";
-		ammo = "Ar2_battery";
-		count = 100;
-	};
-};
 
+#include "CfgMagazines.hpp"
 #include "CfgVehicles.hpp"
 
 class CfgWeapons {
